+++ qbraid/transforms/qasm3/__init__.py
@@ -9,9 +9,9 @@
 # THERE IS NO WARRANTY for the qBraid-SDK, as per Section 15 of the GPL v3.
 
 """
-Module for appyling transformations to OpenQASM 2 programs.
+Module for appyling transformations to OpenQASM 3 programs.
 
-.. currentmodule:: qbraid.transforms.qasm2
+.. currentmodule:: qbraid.transforms.qasm3
 
 Functions
 ----------
@@ -19,11 +19,9 @@ Functions
 .. autosummary::
    :toctree: ../stubs/
 
-   remove_qasm_barriers
-   unfold_qasm_gate_defs
-   flatten_qasm_program
-   decompose_qasm_qelib1
+   qasm3_braket_pre_process
+   qasm3_braket_post_process
+
 
 """
-from .passes import flatten_qasm_program, remove_qasm_barriers, unfold_qasm_gate_defs
-from .qasm_qelib1 import decompose_qasm_qelib1
+from .compat import qasm3_braket_post_process, qasm3_braket_pre_process
